refactor: move toast extension to common module
