+++ common/src/main/kotlin/kr/ohyung/common/shared/UiState.kt
@@ -1,6 +1,6 @@
 /*
  * Created by Lee Oh Hyoung on 2020/08/06 .. 
  */
-package kr.ohyung.common.marks
+package kr.ohyung.common.shared
 
 interface UiState
