+++ presentation/src/main/kotlin/kr/ohyung/navigation/utility/FragmentHelpers.kt
@@ -3,12 +3,9 @@
  */
 package kr.ohyung.navigation.utility
 
-import android.widget.Toast
-import androidx.annotation.StringRes
 import androidx.fragment.app.Fragment
 import kr.ohyung.common.shared.Action
 import kr.ohyung.navigation.ParentActivity
-import kr.ohyung.navigation.NavigationAction
 
 internal val Fragment.sharedViewModel get() =
     (requireActivity() as ParentActivity).sharedViewModel
@@ -18,9 +15,3 @@ internal val Fragment.parentActivity get() =
 
 fun Fragment.navigate(action: Action) =
     (requireActivity() as ParentActivity).navigate(action)
-
-internal fun Fragment.toast(@StringRes message: Int) =
-    Toast.makeText(context, message, Toast.LENGTH_SHORT).show()
-
-internal fun Fragment.toast(message: String) =
-    Toast.makeText(context, message, Toast.LENGTH_SHORT).show()
